+++ hexital/indicators/supertrend.py
@@ -10,7 +10,7 @@ class Supertrend(Indicator):
 
     indicator_name: str = "Supertrend"
     period: int = 7
-    mulitplier: float = 3.0
+    multiplier: float = 3.0
     input_value: str = "close"
 
     def _generate_name(self) -> str:
@@ -32,14 +32,14 @@ class Supertrend(Indicator):
         )
 
         self._add_managed_indicator(
-            "basic_upper",
+            "st_upper",
             indicators.Managed(
                 candles=self.candles,
                 fullname_override="ST_Upper",
             ),
         )
         self._add_managed_indicator(
-            "basic_lower",
+            "st_lower",
             indicators.Managed(
                 candles=self.candles,
                 fullname_override="ST_Lower",
@@ -48,28 +48,27 @@ class Supertrend(Indicator):
 
     def _calculate_reading(self, index: int) -> float | dict | None:
         direction = 1
-        trend = None
 
         if self.reading(f"{self.indicator_name}_atr"):
-            matr = self.mulitplier * self.reading(f"{self.indicator_name}_atr")
+            mid_atr = self.multiplier * self.reading(f"{self.indicator_name}_atr")
 
-            upper = self.reading(f"{self.indicator_name}_HL") + matr
-            lower = self.reading(f"{self.indicator_name}_HL") - matr
+            upper = self.reading(f"{self.indicator_name}_HL") + mid_atr
+            lower = self.reading(f"{self.indicator_name}_HL") - mid_atr
 
-            if self.prev_reading(f"{self.name}.long"):
+            if self.prev_reading("ST_Lower"):
                 if self.reading("close") > self.prev_reading("ST_Upper"):
                     direction = 1
                 elif self.reading("close") < self.prev_reading("ST_Lower"):
                     direction = -1
                 else:
                     direction = self.prev_reading(f"{self.name}.direction")
-                    if direction > 0 and lower < self.prev_reading("ST_Lower"):
+                    if direction == 1 and lower < self.prev_reading("ST_Lower"):
                         lower = self.prev_reading("ST_Lower")
-                    if direction < 0 and upper > self.prev_reading("ST_Upper"):
+                    if direction == -1 and upper > self.prev_reading("ST_Upper"):
                         upper = self.prev_reading("ST_Upper")
 
-            self._managed_indictor("basic_upper").set_reading(upper)
-            self._managed_indictor("basic_lower").set_reading(lower)
+            self._managed_indictor("st_upper").set_reading(upper)
+            self._managed_indictor("st_lower").set_reading(lower)
 
             return {
                 "trend": lower if direction == 1 else upper,
@@ -78,4 +77,4 @@ class Supertrend(Indicator):
                 "short": upper if direction == -1 else None,
             }
 
-        return {"trend": trend, "direction": direction, "long": None, "short": None}
+        return {"trend": None, "direction": direction, "long": None, "short": None}